Split create_plots into 2 scripts
